+++ src/main/java/com/besson/arknights/block/fantastic/PottedConiferPlant.java
@@ -1,0 +1,21 @@
+package com.besson.arknights.block.fantastic;
+
+import com.besson.arknights.block.FurnitureHorizontalFacingBlock;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.shapes.CollisionContext;
+import net.minecraft.world.phys.shapes.VoxelShape;
+
+public class PottedConiferPlant extends FurnitureHorizontalFacingBlock {
+    public PottedConiferPlant(Properties settings) {
+        super(settings);
+    }
+    private static final VoxelShape SHAPE = Block.box(5.25, 0, 5.25, 10.75, 22.5, 10.75);
+
+    @Override
+    public VoxelShape getShape(BlockState pState, BlockGetter pLevel, BlockPos pPos, CollisionContext pContext) {
+        return SHAPE;
+    }
+}