+++ src/main/java/com/besson/arknights/block/fantastic/WideGrayCarpet.java
@@ -1,0 +1,120 @@
+package com.besson.arknights.block.fantastic;
+
+import com.besson.arknights.block.FurnitureHorizontalFacingBlock;
+import com.besson.arknights.block.ModBlocks;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.context.BlockPlaceContext;
+import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.LevelAccessor;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.shapes.CollisionContext;
+import net.minecraft.world.phys.shapes.VoxelShape;
+import org.jetbrains.annotations.Nullable;
+
+public class WideGrayCarpet extends FurnitureHorizontalFacingBlock {
+    public static final VoxelShape SHAPE = Block.box(-16, 0, -16, 32, 0.45, 32);
+    public WideGrayCarpet(Properties settings) {
+        super(settings);
+    }
+
+    @Override
+    public VoxelShape getShape(BlockState pState, BlockGetter pLevel, BlockPos pPos, CollisionContext pContext) {
+        return SHAPE;
+    }
+
+    @Override
+    public void setPlacedBy(Level world, BlockPos pos, BlockState state, @Nullable LivingEntity placer, ItemStack itemStack) {
+        super.setPlacedBy(world, pos, state, placer, itemStack);
+        if (world.isClientSide()){
+            return;
+        }
+        Direction dir = state.getValue(FACING);
+        BlockPos right = pos.relative(dir.getCounterClockWise(),3);
+
+        if (!world.getBlockState(right).isAir()){
+            return;
+        }
+        world.setBlock(right, ModBlocks.WIDE_GRAY_CARPET_RIGHT.get().defaultBlockState().setValue(FACING, dir),3);
+    }
+
+    @Override
+    public void onRemove(BlockState state, Level world, BlockPos pos, BlockState newState, boolean moved) {
+        if (state.is(newState.getBlock())){
+            super.onRemove(state, world, pos, newState, moved);
+            return;
+        }
+
+        BlockPos mainPos = findMainPos(state, pos);
+        if (mainPos == null){
+            return;
+        }
+        Direction dir = state.getValue(FACING);
+        removeAllParts(world, mainPos, dir);
+
+        super.onRemove(state, world, pos, newState, moved);
+    }
+
+    @Override
+    public @Nullable BlockState getStateForPlacement(BlockPlaceContext ctx) {
+        Level world = ctx.getLevel();
+        BlockPos pos = ctx.getClickedPos();
+        BlockState state = super.getStateForPlacement(ctx);
+
+        if (state == null){
+            return null;
+        }
+        Direction dir = state.getValue(FACING);
+        BlockPos right = pos.relative(dir.getCounterClockWise(), 3);
+
+        boolean placeable = world.getBlockState(right).isAir();
+        return placeable ? state : null;
+    }
+
+    @Override
+    public BlockState updateShape(BlockState state, Direction direction, BlockState neighborState, LevelAccessor world, BlockPos pos, BlockPos neighborPos) {
+        if (neighborState.isAir()){
+            return state;
+        }
+        BlockPos mainPos = findMainPos(state, pos);
+        if (mainPos == null){
+            if (world.getBlockState(neighborPos).isAir()){
+                return Blocks.AIR.defaultBlockState();
+            }
+        }
+        return state.setValue(FACING, state.getValue(FACING));
+    }
+
+    private void removeAllParts(Level world, BlockPos mainPos, Direction dir){
+        BlockPos right = mainPos.relative(dir.getCounterClockWise(), 3);
+
+        BlockPos[] pos = {mainPos, right};
+
+        for (BlockPos p : pos){
+            if (isBookcase(world.getBlockState(p))){
+                world.removeBlock(p, false);
+            }
+        }
+    }
+
+    private BlockPos findMainPos(BlockState state, BlockPos pos){
+        if (state.is(ModBlocks.WIDE_GRAY_CARPET_LEFT.get())){
+            return pos;
+        }
+        Direction dir = state.getValue(FACING);
+        if (state.is(ModBlocks.WIDE_GRAY_CARPET_RIGHT.get())){
+            return pos.relative(dir.getClockWise(), 3);
+        }
+        return null;
+    }
+    private boolean isBookcase(BlockState state){
+        return state.is(this) ||
+                state.is(ModBlocks.WIDE_GRAY_CARPET_LEFT.get()) ||
+                state.is(ModBlocks.WIDE_GRAY_CARPET_RIGHT.get());
+    }
+}